+++ packages/react-native-sdk/__tests__/components/CallParticipants.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { mockClientWithUser } from '../mocks/client';
 import { SfuModels } from '@stream-io/video-client';
 import mockParticipant from '../mocks/participant';
-import { A11yComponents } from '../../src/constants/A11yLabels';
+import { ComponentTestIds } from '../../src/constants/TestIds';
 import { mockCall } from '../mocks/call';
 import { act, render, screen, within } from '../utils/RNTLTools';
 import { CallContent } from '../../src/components';
@@ -21,8 +21,8 @@ enum P_IDS {
 const simulateOnViewableItemsChanged = async (
   viewableItems: Array<ViewToken>,
 ) => {
-  const flatList = await screen.findByLabelText(
-    A11yComponents.CALL_PARTICIPANTS_LIST,
+  const flatList = await screen.findByTestId(
+    ComponentTestIds.CALL_PARTICIPANTS_LIST,
   );
   await act(() => {
     flatList.props.onViewableItemsChanged({
@@ -55,28 +55,28 @@ describe('CallParticipants', () => {
     });
 
     expect(
-      await screen.findByLabelText(A11yComponents.CALL_PARTICIPANTS_GRID),
+      await screen.findByTestId(ComponentTestIds.CALL_PARTICIPANTS_GRID),
     ).toBeVisible();
 
     // Locating and verifying that all ParticipantViews are rendered
     const localParticipant = within(
-      screen.getByLabelText(A11yComponents.LOCAL_PARTICIPANT),
+      screen.getByTestId(ComponentTestIds.LOCAL_PARTICIPANT),
     );
     const participant1 = within(
-      screen.getByLabelText(`participant-${P_IDS.REMOTE_1}`),
+      screen.getByTestId(`participant-${P_IDS.REMOTE_1}`),
     );
 
     expect(
-      localParticipant.getByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      localParticipant.getByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveProp('streamURL', 'video-test-url');
 
     expect(
-      participant1.getByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      participant1.getByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveProp('streamURL', 'audio-test-url');
 
     // flat list should not be rendered for 2 participants as we should not wrap them in a grid
     expect(
-      screen.queryByLabelText(A11yComponents.CALL_PARTICIPANTS_LIST),
+      screen.queryByLabelText(ComponentTestIds.CALL_PARTICIPANTS_LIST),
     ).toBeNull();
   });
   it('should render call participants component with spotlight mode with 2 participants', async () => {
@@ -105,12 +105,12 @@ describe('CallParticipants', () => {
     });
 
     expect(
-      await screen.findByLabelText(A11yComponents.CALL_PARTICIPANTS_SPOTLIGHT),
+      await screen.findByTestId(ComponentTestIds.CALL_PARTICIPANTS_SPOTLIGHT),
     ).toBeVisible();
 
     // Since it has a screen share and thereby spotlight, we should render the flatlist even with 2 participants
     expect(
-      await screen.findByLabelText(A11yComponents.CALL_PARTICIPANTS_LIST),
+      await screen.findByTestId(ComponentTestIds.CALL_PARTICIPANTS_LIST),
     ).toBeVisible();
   });
 
@@ -156,35 +156,35 @@ describe('CallParticipants', () => {
 
     // Locating and verifying that all ParticipantViews are rendered
     const localParticipant = within(
-      screen.getByLabelText(`participant-${P_IDS.LOCAL_1}`),
+      screen.getByTestId(`participant-${P_IDS.LOCAL_1}`),
     );
     const participant1 = within(
-      screen.getByLabelText(`participant-${P_IDS.REMOTE_1}`),
+      screen.getByTestId(`participant-${P_IDS.REMOTE_1}`),
     );
     const participant2 = within(
-      screen.getByLabelText(`participant-${P_IDS.REMOTE_2}`),
+      screen.getByTestId(`participant-${P_IDS.REMOTE_2}`),
     );
 
     const participant3 = within(
-      screen.getByLabelText(`participant-${P_IDS.REMOTE_3}`),
+      screen.getByTestId(`participant-${P_IDS.REMOTE_3}`),
     );
 
     // Verifying that the local partic.'s video/audio are rendered within their respective participant
     expect(
-      localParticipant.getByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      localParticipant.getByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveProp('streamURL', 'video-test-url');
     expect(
-      participant1.getByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      participant1.getByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveProp('streamURL', 'audio-test-url');
     expect(
-      participant2.getByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      participant2.getByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveProp('streamURL', 'video-test-url');
     expect(
-      participant3.getByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      participant3.getByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveProp('streamURL', 'video-test-url');
     // Verifying no extra/unknown RTCViews are rendered
     expect(
-      screen.getAllByLabelText(A11yComponents.PARTICIPANT_MEDIA_STREAM),
+      screen.getAllByTestId(ComponentTestIds.PARTICIPANT_MEDIA_STREAM),
     ).toHaveLength(4);
   });
 });