+++ packages/react-native-sdk/__tests__/components/CallParticipantsGrid.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { mockClientWithUser } from '../mocks/client';
 import { SfuModels } from '@stream-io/video-client';
 import mockParticipant from '../mocks/participant';
-import { A11yComponents } from '../../src/constants/A11yLabels';
+import { ComponentTestIds } from '../../src/constants/TestIds';
 import { mockCall } from '../mocks/call';
 import { render, screen } from '../utils/RNTLTools';
 import { CallParticipantsGrid } from '../../src/components/call/internal/CallParticipantsGrid';
@@ -33,7 +33,7 @@ describe('CallParticipantsGrid', () => {
     });
 
     expect(
-      await screen.findByLabelText(A11yComponents.LOCAL_PARTICIPANT_FULLSCREEN),
+      await screen.findByTestId(ComponentTestIds.LOCAL_PARTICIPANT_FULLSCREEN),
     ).toBeVisible();
   });
 });